Add simple e820 memory map
Adds a crappy and simple e820 memory map. It's quite
sloppy and basically the realmode assembly isn't extendable
in the current form.

Also a new make target was added for enabling BIOS debugging
using SeaBIOS. You can clone that repo and build a copy of the
BIOS and the output will be spit out onto the serial console.